refactor: :art: Reorganize data loading by where it is used
Related to #60
